CTX-5357: Added blue color to all prompts, still improving UI. Saving changes
